fix: Update data types for quantity and total profit in ProfitableListing DTO and view to improve accuracy

+++ api/TornTools.Core/DataTransferObjects/ProfitableListingDto.cs
@@ -11,13 +11,13 @@ public class ProfitableListingDto
 
   public long? TornCityMinPrice { get; set; }
   public long? TornCityMaxPrice { get; set; }
-  public int? TornCityQuantity { get; set; }
-  public long? TornCityTotalProfit { get; set; }
+  public long? TornCityQuantity { get; set; }
+  public decimal? TornCityTotalProfit { get; set; }
 
   public long? TornBazaarMinPrice { get; set; }
   public long? TornBazaarMaxPrice { get; set; }
-  public int? TornBazaarQuantity { get; set; }
-  public long? TornBazaarTotalProfit { get; set; }
+  public long? TornBazaarQuantity { get; set; }
+  public decimal? TornBazaarTotalProfit { get; set; }
 
   public DateTimeOffset? TornLastUpdated { get; set; }
 
@@ -25,18 +25,18 @@ public class ProfitableListingDto
 
   public long? Weav3rCityMinPrice { get; set; }
   public long? Weav3rCityMaxPrice { get; set; }
-  public int? Weav3rCityQuantity { get; set; }
-  public long? Weav3rCityTotalProfit { get; set; }
+  public long? Weav3rCityQuantity { get; set; }
+  public decimal? Weav3rCityTotalProfit { get; set; }
 
   public long? Weav3rMarketMinPrice { get; set; }
   public long? Weav3rMarketMaxPrice { get; set; }
-  public int? Weav3rMarketQuantity { get; set; }
-  public long? Weav3rMarketTotalProfit { get; set; }
+  public long? Weav3rMarketQuantity { get; set; }
+  public decimal? Weav3rMarketTotalProfit { get; set; }
 
   public long? Weav3rAnonMinPrice { get; set; }
   public long? Weav3rAnonMaxPrice { get; set; }
-  public int? Weav3rAnonQuantity { get; set; }
-  public long? Weav3rAnonTotalProfit { get; set; }
+  public long? Weav3rAnonQuantity { get; set; }
+  public decimal? Weav3rAnonTotalProfit { get; set; }
 
   public DateTimeOffset? Weav3rLastUpdated { get; set; }
 }
